requirements.txt
google-genai>=1.0.0
gradio>=5.0.0
google-cloud-storage
google-cloud-texttospeech
Pillow
python-dotenv
pydub
tenacity